+++ tests/unit2/test_sprite_scale.py
@@ -35,3 +35,4 @@ def test_sprite():
     window = MyTestWindow(SCREEN_WIDTH, SCREEN_HEIGHT, "Test Text")
     window.test()
     window.close()
+    arcade.cleanup_texture_cache()